more auth tests
